+++ src/Linna/Storage/MysqliStorage.php
@@ -19,52 +19,20 @@ use mysqli;
 class MysqliStorage implements StorageInterface
 {
     /**
-     * @var string Mysql host es. 127.0.0.1
+     * @var array Mysqli connection options
      */
-    protected $host;
-
-    /**
-     * @var string Username for data base connection
-     */
-    protected $user;
-
-    /**
-     * @var string Password for data base connection
-     */
-    protected $password;
-
-    /**
-     * @var string Database name
-     */
-    protected $database;
-
-    /**
-     * @var int Mysql tcp port es. 3306
-     */
-    protected $port;
-
+    protected $options;
+    
     /**
      * Constructor.
      *
-     * @param string $host
-     * @param string $user
-     * @param string $password
-     * @param string $database
-     * @param int    $port
+     * @param array  $options
      */
-    public function __construct(string $host, string $user, string $password, string $database, int $port)
+    public function __construct(array $options)
     {
-        $this->host = $host;
-
-        $this->user = $user;
-
-        $this->password = $password;
-
-        $this->database = $database;
-
-        $this->port = $port;
+        $this->options = $options;
     }
-
+    
     /**
      * Get Resource.
      *
@@ -74,6 +42,13 @@ class MysqliStorage implements StorageInterface
     {
         mysqli_report(MYSQLI_REPORT_ALL);
 
-        return new mysqli($this->host, $this->user, $this->password, $this->database, $this->port);
+        //return new mysqli($this->host, $this->user, $this->password, $this->database, $this->port);
+        return new mysqli(
+            $this->options['host'], 
+            $this->options['user'], 
+            $this->options['password'], 
+            $this->options['database'], 
+            $this->options['port']
+        );
     }
 }